backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
av==10.0.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
ctranslate2==3.21.0
distro==1.8.0
faster-whisper==0.9.0
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
fsspec==2023.10.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
numpy==1.26.2
onnxruntime==1.16.3
openai==1.3.4
packaging==23.2
protobuf==4.25.1
psycopg==3.1.13
psycopg-binary==3.1.13
pydantic==2.5.1
pydantic_core==2.14.3
python-dotenv==1.0.0
pytube==15.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
sympy==1.12
tokenizers==0.14.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
